module3-task1 (first try):  - Added routers, views, layouts.  - Pizza builder logic moved to views/Index.vue
