+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "vue-router": "^3.5.3"
+  }
+}
