+++ src/frontend/src/App.vue
@@ -1,15 +1,17 @@
 <template>
   <div id="app">
-    <Index />
+    <AppLayout>
+      <router-view />
+    </AppLayout>
   </div>
 </template>
 
 <script>
-import Index from "./views/Index.vue";
+import AppLayout from "@/layouts/AppLayout.vue";
 export default {
   name: "App",
   components: {
-    Index,
+    AppLayout,
   },
 };
 </script>
