+++ src/frontend/src/layouts/AppLayoutDefault.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>
+    <slot />
+  </div>
+</template>
+
+<script>
+export default {
+  name: "AppLayoutDefault",
+};
+</script>
