+++ src/frontend/src/layouts/AppLayoutHeader.vue
@@ -1,21 +1,22 @@
 <template>
   <header class="header">
-    <p>What</p>
     <div class="header__logo">
-      <a href="index.html" class="logo">
+      <router-link to="/" class="logo">
         <img
           src="img/logo.svg"
           alt="V!U!E! Pizza logo"
           width="90"
           height="40"
         />
-      </a>
+      </router-link>
     </div>
     <div class="header__cart">
-      <a href="cart.html">{{ cartPrice }} ₽</a>
+      <router-link to="/cart"> {{ cartPrice }} ₽</router-link>
     </div>
     <div class="header__user">
-      <a href="#" class="header__login"><span>Войти</span></a>
+      <router-link to="/login" class="header__login">
+        <span>Войти</span>
+      </router-link>
     </div>
   </header>
 </template>
