+++ src/frontend/src/views/Cart.vue
@@ -1,0 +1,13 @@
+<template>
+  <body>
+    <h1>Cart</h1>
+  </body>
+</template>
+
+<script>
+export default {
+  name: "Cart",
+};
+</script>
+
+<style scoped></style>
