+++ src/frontend/src/views/Login.vue
@@ -1,0 +1,13 @@
+<template>
+  <body>
+    <h1>Login</h1>
+  </body>
+</template>
+
+<script>
+export default {
+  name: "Login",
+};
+</script>
+
+<style scoped></style>
