+++ src/frontend/src/views/Orders.vue
@@ -1,0 +1,13 @@
+<template>
+  <body>
+    <h1>Orders</h1>
+  </body>
+</template>
+
+<script>
+export default {
+  name: "Orders",
+};
+</script>
+
+<style scoped></style>
