+++ src/frontend/src/views/Profile.vue
@@ -1,0 +1,13 @@
+<template>
+  <body>
+    <h1>Profile</h1>
+  </body>
+</template>
+
+<script>
+export default {
+  name: "Profile",
+};
+</script>
+
+<style scoped></style>
